requirements.txt
requests
python-dotenv
finnhub-python